Fix isIP: closes #2972, move integration tests to unit tests and introduce utils.sliceArgs helper
