+++ lib/hooks.js
@@ -89,7 +89,7 @@ var Hooks = {
   runHooks: function(hooks) {
     var self = this
       , fn
-      , fnArgs = Array.prototype.slice.call(arguments, 1)
+      , fnArgs = Utils.sliceArgs(arguments, 1)
       , hookType;
 
     if (typeof fnArgs[fnArgs.length - 1] === 'function') {
