+++ lib/instance-validator.js
@@ -343,11 +343,16 @@ InstanceValidator.prototype._invokeBuiltinValidator = Promise.method(function(va
     'Validation ' + validatorType + ' failed';
 
   if (!Array.isArray(validatorArgs)) {
-    validatorArgs = [validatorArgs];
+    if (validatorType === 'isImmutable') {
+      validatorArgs = [validatorArgs, field];
+    } else if (validatorType === 'isIP') {
+      validatorArgs = [];
+    } else {
+      validatorArgs = [validatorArgs];
+    }
   } else {
     validatorArgs = validatorArgs.slice(0);
   }
-  validatorArgs.push(field);
 
   if (!Validator[validatorType].apply(Validator, [value].concat(validatorArgs))) {
     throw errorMessage;
