+++ lib/sequelize.js
@@ -936,7 +936,7 @@ module.exports = (function() {
    * @return {Sequelize.fn}
    */
   Sequelize.fn = Sequelize.prototype.fn = function(fn) {
-    return new Utils.fn(fn, Array.prototype.slice.call(arguments, 1));
+    return new Utils.fn(fn, Utils.sliceArgs(arguments, 1));
   };
 
   /**
@@ -989,7 +989,7 @@ module.exports = (function() {
    * @return {Sequelize.and}
    */
   Sequelize.and = Sequelize.prototype.and = function() {
-    return new Utils.and(Array.prototype.slice.call(arguments));
+    return new Utils.and(Utils.sliceArgs(arguments));
   };
 
   /**
@@ -1002,7 +1002,7 @@ module.exports = (function() {
    * @return {Sequelize.or}
    */
   Sequelize.or = Sequelize.prototype.or = function() {
-    return new Utils.or(Array.prototype.slice.call(arguments));
+    return new Utils.or(Utils.sliceArgs(arguments));
   };
 
   /**
@@ -1136,7 +1136,7 @@ module.exports = (function() {
   };
 
   Sequelize.prototype.log = function() {
-    var args = [].slice.call(arguments)
+    var args = Utils.sliceArgs(arguments)
       , last = Utils._.last(args)
       , options;
