+++ lib/utils.js
@@ -304,6 +304,15 @@ var Utils = module.exports = {
     return errStack;
   },
 
+  sliceArgs: function (args, begin) {
+    begin = begin || 0;
+    var tmp = new Array(args.length - begin);
+    for (var i = begin; i < args.length; ++i) {
+        tmp[i - begin] = args[i];
+    }
+    return tmp;
+  },
+
   now: function(dialect) {
     var now = new Date();
     if (dialect !== 'postgres') now.setMilliseconds(0);
@@ -339,7 +348,7 @@ var Utils = module.exports = {
 
   col: function(col) {
     if (arguments.length > 1) {
-      col = Array.prototype.slice.call(arguments);
+      col = this.sliceArgs(arguments);
     }
     this.col = col;
   },
